Adds full login/signup functionality

+++ controllers/fruitsController.js
@@ -7,7 +7,13 @@ const router = express.Router();
 
 // Index Route - Retrieve many/all fruits
 router.get('/', (req, res) => {
-  db.Fruit.find({}, (err, allFruits) => {
+  // If the user is not logged in, send them to the login page
+  if (!req.session.currentUser) {
+    return res.redirect('/login');
+  }
+
+  // Find all the fruits that match the id for the logged user
+  db.Fruit.find({ user: req.session.currentUser._id }, (err, allFruits) => {
     if(err) return console.log(err);
 
     res.render('index.ejs', { allFruits: allFruits })
@@ -17,12 +23,21 @@ router.get('/', (req, res) => {
 
 // New Route - Retrieve a form that can be used to create a new fruit
 router.get('/new', (req, res) => {
+  // if user is not logged in, send them to the login page
+  if (!req.session.currentUser) {
+    return res.redirect('/login');
+  }
+
   res.render('new.ejs')
 })
 
 
 // Show Route - Retrieve one fruit
 router.get('/:fruitId', (req, res) => {
+  if (!req.session.currentUser) {
+    return res.redirect('/login');
+  }
+
   db.Fruit.findById(req.params.fruitId, (err, foundFruit) => {
     if(err) return console.log(err);
 
@@ -33,6 +48,14 @@ router.get('/:fruitId', (req, res) => {
 
 // Create Route - Send data to create a new fruit
 router.post('/', (req, res) => {
+
+  if (!req.session.currentUser) {
+    return res.redirect('/login');
+  }
+
+  // The user's information is stored inside of req.session
+  console.log(req.session);
+
   // 1. Convert the data to the correct format
   if (req.body.readyToEat === 'on') {
     req.body.readyToEat = true;
@@ -40,9 +63,17 @@ router.post('/', (req, res) => {
     req.body.readyToEat = false;
   }
 
+  console.log('req.body ==>', req.body);
+
+  // Getting the user's id stored in the cookie
+  // adding to the req.body object
+  req.body.user = req.session.currentUser._id;
+
   // 2. Add that new fruit data into our database
   db.Fruit.create(req.body, (err, createdFruit) => {
     if (err) return console.log(err);
+
+    console.log(createdFruit);
 
     // 3. Redirect back to fruits index.
     res.redirect('/fruits')
@@ -52,6 +83,10 @@ router.post('/', (req, res) => {
 
 // Delete Fruit Route - Delete a fruit from the database
 router.delete('/:fruitId', (req, res) => {
+  if (!req.session.currentUser) {
+    return res.redirect('/login');
+  }
+  
   db.Fruit.findByIdAndDelete(req.params.fruitId, (err) => {
     if(err) return console.log(err);
